Move more Three.js stuff into three folder, fixes sign-renderer tests
